% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{runAppSK}
\alias{runAppSK}
\title{Launch a webpage that shows the alluvial visualization}
\usage{
runAppSK(zTbl, port = 3343, ...)
}
\arguments{
\item{zTbl}{timestamped, sessionized table in proper format}

\item{...}{Additional arguments passed to shiny::runApp}
}
\value{
None
}
\description{
Launch a webpage that shows the alluvial visualization
}
\examples{
runAppSK(sessionTbl)

}
