% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sessionTbl}
\alias{sessionTbl}
\title{Sample timestamped, sessionized dataset to visualize using the alluvial diagram}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 100000 rows and 8 columns.}
\usage{
sessionTbl
}
\description{
Sample timestamped, sessionized dataset to visualize using the alluvial diagram
}
\keyword{datasets}
